added controller.service file for the controller functions
